[1.0.2] remove compile warning

+++ Gamnet/Database/MySQL/Connection.cpp
@@ -71,7 +71,7 @@ namespace Gamnet {	namespace Database { namespace MySQL {
 		}
 		else
 		{
-			impl->affectedRowCount_ = mysql_affected_rows(&conn_);
+			impl->affectedRowCount_ = (unsigned int)mysql_affected_rows(&conn_);
 			impl->lastInsertID_ = static_cast<unsigned int>(mysql_insert_id(&conn_));
 		}
 		return impl;
